feat(i18n): restructure locale constants and add 404 Not Found route
